+++ grails-app/services/com/mini/asaas/address/AddressService.groovy
@@ -1,6 +1,6 @@
-package com.mini.asaas.adress
+package com.mini.asaas.address
 
-import com.mini.asaas.adress.Address
+
 import com.mini.asaas.dto.AddressDTO
 
 import grails.gorm.transactions.Transactional
@@ -11,7 +11,7 @@ import grails.compiler.GrailsCompileStatic
 class AddressService {
 
     public Address save(AddressDTO addressDTO) {
-        Address address = new Address() 
+        Address address = new Address()
         address.cep = addressDTO.cep
         address.city = addressDTO.city
         address.state = addressDTO.state
@@ -23,11 +23,11 @@ class AddressService {
     }
 
     public Address update(AddressDTO addressDTO, Long addressId) {
-        Address address = Address.where{
-            id == addressId 
-            && deleted == false
+        Address address = Address.where {
+            id == addressId
+                    && deleted == false
         }.first()
-        
+
         address.cep = addressDTO.cep
         address.city = addressDTO.city
         address.state = addressDTO.state
@@ -35,7 +35,7 @@ class AddressService {
         address.street = addressDTO.street
         address.number = addressDTO.number
         address.complement = addressDTO.complement
-        
+
         return address.save(failOnError: true)
     }
 }
