+++ grails-app/services/com/mini/asaas/payer/PayerService.groovy
@@ -1,8 +1,7 @@
 package com.mini.asaas.payer
 
-import com.mini.asaas.adress.Address
-import com.mini.asaas.adress.AddressService
-import com.mini.asaas.payer.Payer
+
+import com.mini.asaas.address.AddressService
 import com.mini.asaas.customer.Customer
 import com.mini.asaas.dto.payer.PayerDTO
 import com.mini.asaas.utils.CpfCnpjUtils
@@ -14,7 +13,7 @@ import grails.compiler.GrailsCompileStatic
 @GrailsCompileStatic
 @Transactional
 class PayerService {
-    
+
     AddressService addressService
 
     public Payer save(PayerDTO payerDTO, Long customerId) {
@@ -27,28 +26,28 @@ class PayerService {
         validatedPayer.cpfCnpj = payerDTO.cpfCnpj
         validatedPayer.customer = Customer.get(customerId)
         validatedPayer.personType = CpfCnpjUtils.getPersonType(validatedPayer.cpfCnpj)
-        
+
         validatedPayer.address = addressService.save(payerDTO.addressDTO)
-        
+
         return validatedPayer.save(failOnError: true)
     }
 
     public Payer update(PayerDTO payerDTO, Long payerId) {
-        Payer payer = Payer.where{
-            id == payerId 
-            && deleted == false
+        Payer payer = Payer.where {
+            id == payerId
+                    && deleted == false
         }.first()
 
         Payer validatedPayer = validateSave(payerDTO, payer)
 
         if (validatedPayer.hasErrors()) throw new ValidationException("Erro ao editar pagador", validatedPayer.errors)
-        
+
         validatedPayer.name = payerDTO.name
         validatedPayer.email = payerDTO.email
         validatedPayer.cpfCnpj = payerDTO.cpfCnpj
         validatedPayer.personType = CpfCnpjUtils.getPersonType(validatedPayer.cpfCnpj)
         validatedPayer.address = addressService.update(payerDTO.addressDTO, validatedPayer.address.id)
-        
+
         return validatedPayer.save(failOnError: true)
     }
 
@@ -56,7 +55,7 @@ class PayerService {
         if (!CpfCnpjUtils.validate(payerDTO.cpfCnpj)) {
             payer.errors.reject("cpfCnpj", null, "CPF ou CNPJ inválido.")
         }
-        
+
         return payer
     }
 }
